backend/pyproject.toml
[project]
name = "resume-ragbot-backend"
version = "0.1.0"
description = "Rag-based chatbot for resume Q&A"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "anthropic>=0.76.0",
  "faiss-cpu>=1.13.2",
  "fastapi>=0.128.0",
  "google-genai>=1.59.0",
  "httpx>=0.28.1",
  "odfpy>=1.4.1",
  "openai>=2.15.0",
  "pydantic>=2.12.5",
  "pydantic-settings>=2.12.0",
  "pypdf>=6.6.0",
  "python-docx>=1.2.0",
  "python-multipart>=0.0.21",
  "sentence-transformers>=5.2.0",
  "uvicorn[standard]>=0.40.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/resume_ragbot"]

[dependency-groups]
dev = ["ipython>=9.9.0", "pytest>=9.0.2"]

[tool.pyright]
stubsPath = ["typings"]

[tool.pytest.ini_options]
filterwarnings = ["ignore:builtin type Swig:DeprecationWarning"]
